Move traits to separate header
