+++ rdmalib/include/rdmalib/buffer.hpp
@@ -7,48 +7,16 @@
 
 #include <cereal/cereal.hpp>
 
-// #ifdef USE_LIBFABRIC
 #include <rdma/fabric.h>
 #include <sys/uio.h>
-// #else
-struct ibv_pd;
-struct ibv_mr;
-struct ibv_sge;
-// #endif
+
+#include <rdmalib/libraries.hpp>
 
 namespace rdmalib
 {
 
   template <typename Derived, typename Library>
   struct ScatterGatherElement;
-
-  struct ibverbs;
-  struct libfabric;
-
-  template <typename Library>
-  struct library_traits;
-
-  template <>
-  struct library_traits<ibverbs>
-  {
-    //using type = ibverbs;
-    typedef ibv_mr *mr_t;
-    typedef ibv_pd *pd_t;
-    typedef uint32_t lkey_t;
-    typedef uint32_t rkey_t;
-    typedef ibv_sge sge_t;
-  };
-
-  template <>
-  struct library_traits<libfabric>
-  {
-    //using type = libfabric;
-    typedef fid_mr *mr_t;
-    typedef fid_domain *pd_t;
-    typedef void *lkey_t;
-    typedef uint64_t rkey_t;
-    typedef iovec sge_t;
-  };
 
   namespace impl
   {
